First version for generation market ids
Make sure you have done the following steps:

1) Got an Appkey associated with your account (google appkey developer betfair and follow instructions)
2) Updated lines 4-7 in the above code for your account details
3) Got a folder in the same folder as your project (PGT_v3/betfair) named ssl_certs (Library/Ravi/PGT_v3/betfair/ssl_certs) with the correct ssl keys, named as your account username.crt ; username.csr ; username.key

Afterwards running the above code should output a series of 4 lists. Each series gives you the name of the two teams playing; the market Ids for each over/under market for that specific match; which x in over/under x goals that market Id refers to; and the amount bet on each market Id... see example below. Note that the 3 lists after the match name are in the same order; so the second number in the market Ids list refers to over/under 4.5 goals in the below example, and there is 131.48£ currently matched on that market.  

Ath Bilbao (W) v Levante (W)
['1.125263351', '1.125263309', '1.125263313', '1.125263308', '1.125263304', '1.125263306']
['Over/Under 7.5 Goals', 'Over/Under 4.5 Goals', 'Over/Under 3.5 Goals', 'Over/Under 2.5 Goals', 'Over/Under 6.5 Goals', 'Over/Under 5.5 Goals']
[164.92, 131.48, 1355.25, 3177.98, 178.38, 150.94]