Create MarketPrices
This file looks at the MarketIds we have filtered for in the Main.py file and fetches the current best prices and volumes available on those prices for back and lay Over x.5 goals and Under x.5 goals. Quite slow for my experience to run the whole thing so we will need to filter it down a bit more and then only send a handful of match requests. This may be on the initial Main program where we can filter for event (i.e Premier league; Euros etc.). You should be able to use this format to get you started anyway.